Revert moving lookup stuff to IdentityHandler
